feat(collapsible): bundle function with the component - check if tagName is undefined
